refactor/BE-01 : auth, users 모듈 분리
