+++ backend/app/auth/router.py
@@ -5,16 +5,14 @@ from sqlalchemy.exc import IntegrityError
 
 from app.core.database import get_db
 from app.core.config import get_settings
-from app.auth.models import User
+from app.users.models import User
 from app.auth.schemas import (
     UserCreate,
     UserLogin,
     UserLogout,
     TokenResponse,
-    UserResponse,
-    BirthDateUpdate,
-    WeeklyTargetUpdate,
 )
+from app.users.schemas import UserResponse
 from app.auth.utils import (
     get_password_hash,
     verify_password,
@@ -85,42 +83,6 @@ def login(user_data: UserLogin, db: Session = Depends(get_db)):
         refresh_token=refresh_token,
         token_type="Bearer",
     )
-
-
-@router.patch("/me/birth-date", response_model=UserResponse)
-def update_birth_date(
-    data: BirthDateUpdate,
-    token_data: tuple[str, str] = Depends(verify_access_token),
-    db: Session = Depends(get_db),
-):
-    email, _ = token_data
-    user = db.query(User).filter(User.email == email).first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="유저를 찾을 수 없습니다."
-        )
-    user.birth_date = data.birth_date  # type: ignore[assignment]
-    db.commit()
-    db.refresh(user)
-    return user
-
-
-@router.patch("/me/weekly-target", response_model=UserResponse)
-def update_weekly_target(
-    data: WeeklyTargetUpdate,
-    token_data: tuple[str, str] = Depends(verify_access_token),
-    db: Session = Depends(get_db),
-):
-    email, _ = token_data
-    user = db.query(User).filter(User.email == email).first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="유저를 찾을 수 없습니다."
-        )
-    user.weekly_target = data.weekly_target  # type: ignore[assignment]
-    db.commit()
-    db.refresh(user)
-    return user
 
 
 @router.post("/logout", status_code=status.HTTP_204_NO_CONTENT)
