+++ backend/app/main.py
@@ -2,6 +2,7 @@ from contextlib import asynccontextmanager
 from fastapi import FastAPI
 from app.core.database import get_engine, Base
 from app.auth.router import router as auth_router
+from app.users.router import router as users_router
 
 
 @asynccontextmanager
@@ -24,3 +25,4 @@ async def health_check():
 
 
 app.include_router(auth_router)
+app.include_router(users_router)
